move swi2 examples to .docs/Notebooks, add descriptions from tm6-a46
